Edited Pine City Mall Files
I made the tabs buttons font size bigger so it would be bigger on a mobile screen because the font was too small on my phone.  I increased the margin-top amounts to accommodate this. I also increased the size of the about store font and shifted the text with an extra div id.